refactor: modularise different parts of the code to make it more scalable in the future
