+++ src/app/_schemas/auth.ts
@@ -10,13 +10,3 @@ export const signInFormSchema = z.object({
     .min(1, 'Password field is required.')
     .min(6, 'Password must be at least 6 characters.'),
 });
-
-export type SignInFormState =
-  | {
-      errors?: {
-        email?: string[];
-        password?: string[];
-      };
-      invalidCredentials?: boolean;
-    }
-  | undefined;
