+++ src/app/_types/auth.ts
@@ -1,0 +1,9 @@
+export type SignInFormState =
+  | {
+      errors?: {
+        email?: string[];
+        password?: string[];
+      };
+      invalidCredentials?: boolean;
+    }
+  | undefined;
